+++ server/helpers/spawnExec.js
@@ -1,9 +1,9 @@
 const { spawn } = require('child_process');
 
-const spawnExec = (reqBody) => new Promise((resolve, reject) => {
+const spawnExec = (imageFilename) => new Promise((resolve, reject) => {
     /**
      */
-    var imageFilename = reqBody;
+    // imageFilename = "plane.png";
     console.log(`output: ${imageFilename}`)
     const ls = spawn('sh network-call.sh',[imageFilename],{ shell: true }); /// this is our command
     // https://nodejs.org/api/child_process.html#child_process_child_process_spawn_command_args_options
